Move macro from tests/reader.rs to helper module `helpers`
Macro moved to helper module because `async-tokio.rs` runs imported `reader.rs` tests,
which is unwanted consequence
